[feat] #16 hurry test 폴더 분리
